rename to nswitch
