+++ core/sdk/src/main/java/com/viam/sdk/core/resource/ResourceManager.java
@@ -55,7 +55,7 @@ import com.viam.sdk.core.component.sensor.SensorRPCService;
 import com.viam.sdk.core.component.servo.Servo;
 import com.viam.sdk.core.component.servo.ServoRPCClient;
 import com.viam.sdk.core.component.servo.ServoRPCService;
-import com.viam.sdk.core.component.switch.*;
+import com.viam.sdk.core.component.nswitch.*;
 import com.viam.sdk.core.exception.DuplicateResourceException;
 import com.viam.sdk.core.exception.ResourceNotFoundException;
 import com.viam.sdk.core.service.datamanager.DataManager;
