+++ core/sdk/src/main/kotlin/com/viam/sdk/core/component/switch/Switch.kt
@@ -1,4 +1,4 @@
-package com.viam.sdk.core.component.switch
+package com.viam.sdk.core.component.nswitch
 
 import com.google.protobuf.Struct
 import com.google.protobuf.Value
