+++ core/sdk/src/main/kotlin/com/viam/sdk/core/component/switch/SwitchRPCClient.kt
@@ -1,12 +1,11 @@
-package com.viam.sdk.core.component.switch
+package com.viam.sdk.core.component.nswitch
 
 import com.google.protobuf.Struct
 import com.google.protobuf.Value
 import com.viam.common.v1.Common
-import com.viam.common.v1.Common.GetGeometriesRequest
-import com.viam.component.switch.v1.SwitchServiceGrpc
-import com.viam.component.switch.v1.SwitchServiceGrpc.SwitchServiceBlockingStub
-import com.viam.component.switch.v1.Switch.*
+import com.viam.component.nswitch.v1.SwitchServiceGrpc
+import com.viam.component.nswitch.v1.SwitchServiceGrpc.SwitchServiceBlockingStub
+import com.viam.component.nswitch.v1.Switch.*
 import com.viam.sdk.core.rpc.Channel
 import java.util.*
 import kotlin.jvm.optionals.getOrDefault
