+++ core/sdk/src/test/kotlin/com/viam/sdk/core/component/switch/SwitchRPCClientTest.kt
@@ -1,4 +1,4 @@
-package com.viam.sdk.core.component.switch
+package com.viam.sdk.core.component.nswitch
 
 import com.google.protobuf.Struct
 import com.google.protobuf.Value
